addressed Enum internal switching code smell
