+++ code/src/test/pcgen/io/freemarker/PCHasVarFunctionTest.java
@@ -67,7 +67,7 @@ public class PCHasVarFunctionTest extends AbstractJunit5CharacterTestCase
 	public void testExec() throws Exception
 	{
 		PlayerCharacter pc = getCharacter();
-		ExportHandler eh = new ExportHandler(new File(""));
+		ExportHandler eh = ExportHandler.createExportHandler(new File(""));
 		PCHasVarFunction pchv = new PCHasVarFunction(pc, eh);
 		
 		Boolean result = (Boolean) pchv.exec(Collections.singletonList("FooV"));
