+++ code/src/test/plugin/exporttokens/CampaignHistoryTokenTest.java
@@ -129,7 +129,7 @@ public class CampaignHistoryTokenTest  extends AbstractCharacterTestCase
 	{
 		StringWriter retWriter = new StringWriter();
 		BufferedWriter bufWriter = new BufferedWriter(retWriter);
-		ExportHandler export = new ExportHandler(new File(""));
+		ExportHandler export = ExportHandler.createExportHandler(new File(""));
 		export.replaceTokenSkipMath(pc, token, bufWriter);
 		retWriter.flush();
